Now in ES6, decomposed.
