+++ README.md
@@ -2,7 +2,7 @@
 
 Richard Lobb, Michelle Hsieh
 
-Version 1.0.2, 8 December 2022.
+Version 1.1.0, 12 December 2022.
 
 Github repo: https://github.com/trampgeek/moodle-filter_ace_inline
 
@@ -283,6 +283,12 @@ Example:
 
 
 ## Change History
+
+ * Version 1.1.0, 12 December 2022.
+    * Entire program is now in Vanilla ES.
+        * Decomposed with UiParameters as a class.
+        * ace_inline_code.js is the entry class.
+        * Remaining supporting classes are in modules.
 
  * Version 1.0.2, 8 December 2022. 
     * Removed all jQuery from code to prepare for ES6 transition.
