+++ lang/en/filter_ace_inline.php
@@ -17,35 +17,38 @@
  * Strings for component 'filter_ace_inline', language 'en'
  *
  * @package    filter_ace_inline
- * @copyright  2017 Richard Jones https://richardnz.net
+ * @copyright  Richard Jones, Michelle Hsieh 2017, 2022
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
 $string['filtername'] = 'Ace inline';
 
 // Settings strings.
-$string['settings_heading'] = 'Ace inline filter settings';
-$string['settings_desc'] = 'Change the settings for this filter.';
+$string['default_button_label'] = 'Try it!';
 $string['settings_button_label'] = 'Button label';
 $string['settings_button_label_desc'] = 'The label for the "Try it!" button that
     the user clicks to run the code';
-$string['settings_dark_theme'] = 'Set when to use dark theme';
-$string['settings_dark_never'] = 'Never';
 $string['settings_dark_always'] = 'Always';
+$string['settings_dark_never'] = 'Never';
 $string['settings_dark_preference'] = 'When OS or browser prefers dark';
+$string['settings_dark_theme'] = 'Set when to use dark theme';
 $string['settings_dark_theme_desc'] = 'Select when to use a dark theme for Ace
 instead of the default light theme. Can be overridden by an individual instance.
 \'sometimes\' behaves according to the browser\'s response to the \'prefers-color-scheme:dark\' media query.';
-$string['default_button_label'] = 'Try it!';
+$string['settings_desc'] = 'Change the settings for this filter.';
+$string['settings_heading'] = 'Ace inline filter settings';
+
+// Error strings.
 $string['error_access_denied'] = 'Sandbox server access denied';
 $string['error_element_unknown'] = 'Id not found for element';
-$string['error_user_params'] = 'User Config Error';
-$string['error_unknown_language'] = 'Unknown language requested';
-$string['error_submission_limit_reached'] = 'Jobe sandbox submission limit reached';
-$string['error_sandbox_server_overload'] = 'Jobe server overload';
-$string['error_jobe_unknown'] = 'Unknown error from Jobe server';
-$string['error_timeout'] = 'Time limit exceeded';
-$string['error_memory_limit'] = 'Memory limit exceeded';
 $string['error_excessive_output'] = 'Excessive output';
+$string['error_jobe_unknown'] = 'Unknown error from Jobe server';
+$string['error_memory_limit'] = 'Memory limit exceeded';
+$string['error_sandbox_server_overload'] = 'Jobe server overload';
+$string['error_script_unknown'] = 'Mapped function was not found on page';
+$string['error_submission_limit_reached'] = 'Jobe sandbox submission limit reached';
+$string['error_timeout'] = 'Time limit exceeded';
+$string['error_unknown_language'] = 'Unknown language requested';
 $string['error_unknown_runtime'] = 'Unknown runtime error';
+$string['error_user_params'] = 'User Config Error';
