+++ tests/behat/codemapper.feature
@@ -28,3 +28,9 @@ Feature: Codemapper functionality allows transformation of input code prior to e
     And I should not see "Yes, this ran Python"
     And I press "Try it!"
     Then I should see "Yes, this ran Python"
+
+  Scenario: Checks error with Codemapper function throws appropriate error
+    When I am on the "codemapperdemo" "core_question > preview" page logged in as teacher
+    And I should not see "Yes, this ran Python"
+    And I press "nothere"
+    Then I should see "Mapped function was not found on page"
